Pull out schemas to associated folder
