+++ data_transfer/schemas/ucam.py
@@ -1,0 +1,48 @@
+
+from datetime import datetime
+from dataclasses import dataclass
+
+
+@dataclass
+class Payload:
+    """
+    UCAM uses different field naming than our middleware. This Payload
+    Creates consistency between how Records are used
+
+    Current return values from UCAM DB:
+        SubjectID, Deviations, DeviceID, 
+        VTTGeneratedParticipantID, StartDate, 
+        EndDate, SubjectGroup
+    """
+    patient_id: str
+    disease: str
+
+    device_id: str
+    devitations: str
+    vttsma_id: str
+    start_wear: datetime
+    end_wear: datetime
+
+
+@dataclass
+class Device:
+    id: str
+    vttsma_id: str
+    devitations: str
+    start_wear: datetime
+    end_wear: datetime
+
+
+@dataclass
+class Patient:
+    id: str
+    disease: str
+
+
+@dataclass
+class PatientRecord:
+    """
+    Simplifies accessing Devices a Patient used.
+    """
+    patient: Patient
+    devices: [Device]
